+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 ### 🔧 Fixed
 
 * Use `>=` when in `dump_targets_to_parquet` when comparing the current MJD with `observe_until_mjd`.
-* Handle crash when there are no ellegible targets to dump.
+* Handle crash when there are no valid targets to dump.
 
 
 ## 1.0.1 - 2025-02-03